// src/types/models.ts
export interface IPool {
    vault: string;
    totalPoolCollateral: string;
    poolCR: string;
    vaultCR: string;
    feeShare: string;
    agentName: string;
    vaultType: string;
    poolExitCR: string;
    poolTopupCR: string;
    freeLots: string;
    status: boolean;
    tokenAddress: string;
    transferableTokens: string;
    fassetDebt: string;
    userPoolBalance?: string;
    userPoolFees?: string;
    userPoolNatBalance?: string;
    userPoolNatBalanceInUSD?: string;
    userPoolShare?: string;
    health: number;
    mintCount: number;
    numLiquidations: number;
    pool: string;
    redeemRate: string;
    collateralToken: string;
    poolCollateralUSD: string;
    url: string;
    vaultCollateral: string;
    nonTimeLocked?: string;
    mintingPoolCR: string;
    poolCCBCR: string;
    poolMinCR: string;
    poolSafetyCR: string;
    mintingVaultCR: string;
    vaultCCBCR: string;
    vaultMinCR: string;
    vaultSafetyCR: string;
    mintFee: string;
    userPoolFeesUSD?: string;
    allLots: number;
    mintedAssets: string;
    mintedUSD: string;
    remainingAssets: string;
    remainingUSD: string;
    poolOnlyCollateralUSD: string;
    vaultOnlyCollateralUSD: string;
    limitUSD: string;
    totalPortfolioValueUSD: string;
    description: string;
    lifetimeClaimedPoolFormatted?: string;
    lifetimeClaimedPoolUSDFormatted?: string;
    infoUrl: string;
}